Unify PS3 input driver
